+++ lagerbestand/js/ui-manager.js
@@ -876,8 +876,11 @@ class UIManager {
 
             this.closeMaterialModal();
 
-            // Refresh materials list if on materials tab
-            this.renderMaterialsList();
+            // Refresh materials list if on materials tab, preserving state
+            this.renderMaterialsList({ 
+                preserveState: true, 
+                highlightMaterialCode: mode === 'edit' ? code : null 
+            });
             
             // Update undo/redo buttons
             if (this.updateUndoRedoButtons) {
@@ -935,7 +938,8 @@ class UIManager {
         }
 
         if (this.dataManager.deleteMaterial(code)) {
-            this.renderMaterialsList();
+            // Preserve state when deleting (but don't highlight anything)
+            this.renderMaterialsList({ preserveState: true });
             
             // Update undo/redo buttons
             if (this.updateUndoRedoButtons) {